requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.6
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
tensorflow==2.15.0
yfinance==0.2.35
python-binance==1.0.19
alpha-vantage==2.3.1
websockets==12.0
httpx==0.26.0
python-dotenv==1.0.0
aioredis==2.0.1
asyncpg==0.29.0
scipy==1.12.0
statsmodels==0.14.1
xgboost==2.0.3
cvxpy==1.4.2
arch==6.3.0
ta-lib==0.4.28
pykalman==0.9.7
hmmlearn==0.3.2
pmdarima==2.0.4
